+++ core/admin.py
@@ -1,9 +1,7 @@
 from django.contrib import admin
 
 from simple_history.admin import SimpleHistoryAdmin
-from .models import Habit, Profile, Day
+from .models import Profile
 
 
-admin.site.register(Habit, SimpleHistoryAdmin)
 admin.site.register(Profile, SimpleHistoryAdmin)
-admin.site.register(Day, SimpleHistoryAdmin)
