+++ habits/admin.py
@@ -1,0 +1,8 @@
+from django.contrib import admin
+
+from simple_history.admin import SimpleHistoryAdmin
+from .models import Habit, Day
+
+
+admin.site.register(Habit, SimpleHistoryAdmin)
+admin.site.register(Day, SimpleHistoryAdmin)
